+++ custom_components/nexo/const.py
@@ -1,3 +1,2 @@
 """Constants for the nexo integration."""
-
 DOMAIN = "nexo"